Class that defines the Changelog Table Row Table in the Generic Report

This R6 class controls one row of the Changelog table that documents the date, version, reason and what changed in the update on the document.

Active bindings

date

Active binding for setting the date of changes

version

Active binding for setting the version of the document

why_update

Active binding for setting the reason for the update

what_changed

Active binding for setting the section and title impacted

Methods

Public methods

ChangelogTableRow$new()

ChangelogTableRow$clone()

Method new()

Usage

ChangelogTableRow$new(
  date = NA_character_,
  version = NA_character_,
  why_update = NA_character_,
  what_changed = NA_character_
)

Arguments

date

String. Date (DD-Mmm-YYYY) when modifications were performed

version

String. Document version

why_update

String. Reason for updating the document

what_changed

String. Sections and titles that where impacted with the modifications

Examples

ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')

Method clone()

The objects of this class are cloneable with this method.

Usage

ChangelogTableRow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `ChangelogTableRow$new`
## ------------------------------------------------

ctr <- ChangelogTableRow$new(
 date = '01-Oct-2024',
 version = 'Initial version',
 why_update = 'Create first version of the document',
 what_changed = 'Document creation')