Create caption yaml

Given a list of captions, it creates a template yaml file with

type. Indicates if it is a table (TBL) or a figure (FIG)

title. The title or caption of the table/figure in the template docx file

file. The name of the file (csv or image) to be used in the new version of the docx

widths. For tables only, a pre-defined column widths (useful for model parameter table, for example) An example of the yaml structure:

occurrence. Int. To be used when captions are duplicated for more than one figure/table. To set the order they appear in the word document.

output_1:
  type: TBL
  title: "Caption of the output 1"
  file: ~
  widths: ~
  occurrence: ~
output_2:
  type: FIG
  title: "Caption of the output 2"
  file: ~
  widths: ~
  occurrence: ~
  

Usage

create_yml(figure_table_captions_sources, yaml_filename)

Arguments

figure_table_captions_sources

List of the tables and figures with its captions

yaml_filename

String. Output yaml filename, e.g. "/path/to/file/filename.yml"

Examples

if (FALSE) { # \dontrun{
captions <- extract_figure_table_captions(system.file("use_cases/02_automated_reporting",
                                         "Automated_Reporting_Example.docx",
                                         package="rdocx"))
create_yml(captions, "./example_yml.docx")
} # }